// src/types/models.ts
// * New game form data interface
export interface NewGameFormData {
  symbol: string;
  faction: string;
};

// * User game access token form interface
export interface GameAccessToken {
  token: string;
};

// * Accept contract with id form interface
export interface AcceptContractId {
  contractId: string;
};